feat(solver): complete rational roots Étape 1 infrastructure
Add `in` truth checker, `substitute` normalizer (renamed from `replace`),
`divisors(n)` Python calculator, structural TermRef equality, and parser
reserved-word fix to prevent keywords from being parsed as identifiers.

- Move `in` symbol definition fully to Rust (op_in.rs) with truth checker
- Rename `replace` → `substitute` across codebase
- Add `divisors.sym` with Python calculator for integer divisors
- Make TermRef::PartialEq structural instead of pointer-based, add same()
- Add reserved_word() negative lookahead in PEG parser for in/is/as
- Fix clippy cmp_owned warnings in plus.rs
- Add design doc (doc/rational_roots_ru.md)

+++ src/solver/term/symbol/base/plus.rs
@@ -123,9 +123,9 @@ fn attach_constant(root: &mut TermMut, constant: Decimal) {
 }
 
 fn merge_mul_const(mut root: TermBuf, d: Decimal) -> TermBuf {
-    if d == Decimal::from(1) {
+    if d == 1 {
         return root;
-    } else if d == Decimal::from(0) {
+    } else if d == 0 {
         return TermBuf::zero();
     }
     let constant = TermBuf::number(d);
